refactor: extract stage execution into execution.py
Move _execute_stage_list and _validate_stage_prerequisites into
execution.py (133 lines). simulate.py drops from 592 to 472 lines,
now within the ~450 target. Update test imports and patch targets.
